requirements.txt
# Robot Framework Core
robotframework>=7.0.1
robotframework-seleniumlibrary>=6.0.0
robotframework-requests>=0.9.0

# Selenium
selenium>=4.20.0
webdriver-manager>=4.0.0

# HTTP Requests
requests>=2.32.0

# Image Processing for Visual Validation
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.22.0

# Configuration
PyYAML>=6.0

# Test Data Generation
Faker>=25.0.0

# Reporting
Jinja2>=3.1.0

# Utilities
python-dotenv>=1.0.0
